+++ src/core/component/state/CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 #### :house: Internal
 
 * Added `experiments`
+* Added `interface/State`
 
 ## v3.0.0-rc.37 (2020-07-20)